Added optional callback function.
This callback function is just a wrapper for the underlying
callback in the NPM request module. It is optional to keep
backward compatibility but can be useful if you need to know the
result of your request sent to Voxbone or the number of fragments
needed to carry the message in case of fragmentation.