Added 'final' parameter to response body.
The callback function for the sendSMS function now contains
the necessary parameters to determine if the SMS has been transferred
completely to Voxbone for delivery.

If set to true, the 'final' parameter of the response body indicates
that the last fragment has been transferred to Voxbone in case of
fragmentation. It is set to true if the SMS does not need to be
fragmented.